requirements.txt
asgiref==3.6.0
Babel==2.11.0
Django==4.1.5
django-phonenumber-field==7.0.2
djangorestframework==3.14.0
phonenumbers==8.13.4
Pillow==9.4.0
psycopg2==2.9.5
python-dotenv==0.21.1
pytz==2022.7.1
sqlparse==0.4.3
